requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pymongo==4.6.1
google-genai==0.3.0
jinja2==3.1.3
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.31.0
